+++ resources/main.js
@@ -4,6 +4,7 @@ let state = vscode.getState();
 if (!state) {
   state = {};
 }
+vscode.postMessage({ command: 'sync', data: state });
 
 for (let el of elements) {
   let name = el.getAttribute("name");
@@ -14,10 +15,12 @@ for (let el of elements) {
       el.value = state[name];
     }
     el.addEventListener('keyup', (e) => {
-      console.log("Keyup");
       if (e.key === 'Enter' || e.keyCode === 13) {
         updateState(name, el.value);
       }
+    });
+    el.addEventListener('blur', (e) => {
+        updateState(name, el.value);
     });
   } else if ((el.type === 'button') || (el.type === 'submit')) {
     el.addEventListener('click', () => vscode.postMessage({ command: name }));
@@ -39,10 +42,11 @@ function updateState(key, value) {
   if (!state) {
     state = {};
   }
+  var oldVal = state[key];
   state[key] = value;
-  vscode.setState(state);
 
-  // Sync with extension
-  console.log(JSON.stringify(state));
-  vscode.postMessage({ command: 'sync', data: state });
+  if (oldVal !== value) {
+    vscode.setState(state);
+    vscode.postMessage({ command: 'sync', data: state });
+  }
 }