+++ CHANGELOG.md
@@ -2,17 +2,25 @@
 
 All notable changes to the `TaskFalcon` extension will be documented in this file.
 
+## 0.8.8
+* Includes TaskFalcon 0.9.17. See [TaskFalcon Release Notes](https://taskfalcon.org/release-notes/index.html).
+* Added support for filter tags
+
+## 0.8.7
+* Includes TaskFalcon 0.9.15. See [TaskFalcon Release Notes](https://taskfalcon.org/release-notes/index.html).
+* Clicking on a task/resource on the preview image will navigate directly to the task/resource definition. Clicking multiple times will navigate through updates of the same tasks/resources.
+
 ## 0.8.6
-* Includes TaskFalcon 0.9.14 See [TaskFalcon Release Notes](https://taskfalcon.org/release-notes/index.html).
+* Includes TaskFalcon 0.9.14. See [TaskFalcon Release Notes](https://taskfalcon.org/release-notes/index.html).
 * Bugfix release
 
 ## 0.8.5
-* Includes TaskFalcon 0.9.13 See [TaskFalcon Release Notes](https://taskfalcon.org/release-notes/index.html).
+* Includes TaskFalcon 0.9.13. See [TaskFalcon Release Notes](https://taskfalcon.org/release-notes/index.html).
 * Improved live syntax validation
 * Added rendering start/end dates, Done-Column and custom parameters
 
 ## 0.8.4
-* Includes TaskFalcon 0.9.12 See [TaskFalcon Release Notes](https://taskfalcon.org/release-notes/index.html).
+* Includes TaskFalcon 0.9.12. See [TaskFalcon Release Notes](https://taskfalcon.org/release-notes/index.html).
 * Improved live syntax validation, added links to documentation
 * Added scaling, scheduling date and update options to live preview